+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Fixes
 - Handle index db corruption and warn in UI (mihran113) 
+- Handle and skip corrupted runs (alberttorosyan)
 
 ## 3.19.3 Apr 17, 2024
 - Resolve issue with new runs after tracking queue shutdown (mihran113)